fix e2e tests (#80)
* Refactor buffer mapping and logging dependencies

Refactored `Buffer.native.kt` to reuse a global memory callback for mapping operations. Replaced SLF4J with Logback for improved logging capabilities. Updated the version of `wgpu4k-native` and implemented the `poll` function in `Device.native.kt`.

* Normalize texture format values and improve generation logic

Updated texture format values to use consistent naming without hyphens. Enhanced the generator logic with `getValueProviderFrom` and a converter function to handle name transformations dynamically based on enumeration type. Removed an unnecessary import and an experimental annotation from the build script.

* Switch from `.name` to `.value` for enum mappings.

This ensures consistent retrieval of enum values across the codebase. The change improves compatibility with the underlying JS/WASM APIs, where `.value` is the expected format for enum handling.

* Fix type inconsistencies in GPUCanvasContext and canvas dimensions

Updated `GPUCanvasContext.canvas` to use a strongly-typed `HtmlElementWithSize` interface. Converted canvas width and height to `UInt` to ensure proper numeric handling and type safety.

* Use `Uint32Array` for dynamicOffsets in setBindGroup.

Updated `setBindGroup` methods in `ComputePassEncoder` and `RenderPassEncoder` to cast or default `dynamicOffsets` to `Uint32Array`. This ensures proper type compliance with WebGPU's expected API behavior.

* Update GPUBuffer size type to JsBigInt for large buffer support

The GPUBuffer interface now uses JsBigInt instead of JsNumber for the `size` property to properly handle larger buffer sizes. Updated the `size` getter in Buffer.wasmJs.kt to accommodate this change while ensuring compatibility with existing functionality.

* Update run configurations for web examples

Replaced the outdated "run web example" configuration with distinct "run web js example" and "run web wasm example" configurations. This provides clearer separation and support for JS and WASM-specific tasks.

* Remove unused function for JsNumber to ULong conversion

The `infer` function and the related `toULong` extension were removed as they were no longer utilized in the codebase. This cleanup improves maintainability by eliminating dead code.

* Update stencil mask types and add UInt to Number mapping

Replaced GPUStencilValue with Number for stencil masks to improve type consistency. Added a new utility function to map UInt values to Number in the mapper package. Updated the RenderPipelineDescriptor to utilize the new mapping for stencil mask values.

+++ wgpu4k/src/wasmJsMain/kotlin/io.ygdrasil.wgpu/Buffer.wasmJs.kt
@@ -3,7 +3,6 @@ package io.ygdrasil.webgpu
 import io.ygdrasil.webgpu.internal.js.GPUBuffer
 import io.ygdrasil.webgpu.internal.js.mapJsArray
 import io.ygdrasil.webgpu.internal.js.toJsNumber
-import io.ygdrasil.webgpu.internal.js.toULong
 import org.khronos.webgl.Float32Array
 import org.khronos.webgl.Int16Array
 import org.khronos.webgl.Int32Array
@@ -13,7 +12,9 @@ import org.khronos.webgl.get
 actual class Buffer(internal val handler: GPUBuffer) : AutoCloseable {
 
     actual val size: GPUSize64
-        get() = handler.size.toULong()
+        get() = handler
+            .size
+            .toLong().toULong()
     actual val usage: Set<BufferUsage>
         get() = BufferUsage.entries.filter { it.value.toUInt() and handler.usage != 0u }.toSet()
     actual val mapState: BufferMapState
